fix: 서버 주소 바뀐 것 반영
