+++ src/api/auth.ts
@@ -66,19 +66,20 @@ export const findPwd = async (userData: any) => {
 
   try {
     const res = await axios({
-      url: `${baseUrl}/member/modify-password/${id}`,
-      data: { password: password },
+      url: `${baseUrl}/member/modify-password`,
+      data: { memberId: id, password: password },
       method: "put",
       headers: {
         "Content-Type": "application/json",
       },
     });
     if (res.status === 200) {
-      alert("비밀번호가 바뀌었습니다.");
+      alert("비밀번호가 변경되었습니다.");
       window.location.replace("/signin");
     }
   } catch (error) {
     alert("비밀번호 변경에 실패했습니다.");
+    console.log(error);
   }
 };
 
@@ -150,26 +151,6 @@ export const getMyData = async (token: string) => {
     console.log(error);
   }
 };
-
-//userId로 정보 가져오기
-//export const getUserDataById = async (userId: string) => {
-//  try {
-//    const res = await axios({
-//      url: `${baseUrl}/member/info/${userId}`,
-//      method: "get",
-//      headers: {
-//        "Content-Type": "application/json",
-//      },
-//    });
-//
-//    if (res.status === 200) {
-//      const user = await res.data;
-//      return user;
-//    }
-//  } catch (error) {
-//    console.log(error);
-//  }
-//};
 
 // 유저 아이디로 정보 불러오기
 export function useGetUserDataById(userId?: string) {
